+++ generate/testdata/queries/mypkg/types.go
@@ -1,0 +1,3 @@
+package mypkg
+
+type ID string